AI/requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
fastapi>=0.115.6
uvicorn>=0.34.0
celery>=5.4.0
redis>=5.0.1